mbc3: added support for mbc3 (w/o rtc)

+++ include/core/cart.h
@@ -30,6 +30,6 @@ namespace Umibozu {
     Cartridge();
     ~Cartridge();
     
-    u8 read8(const u16);
+    u8 read8(const u64);
   };
 }  // namespace Umibozu
